src/api/requirements.txt
flask
flask-restx
paho-mqtt
requests
influxdb-client
